Move socket utility functions to pkg/util/system/socket
