+++ pkg/security/agent/agent.go
@@ -27,11 +27,11 @@ import (
 	"github.com/DataDog/datadog-agent/comp/logs/agent/config"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/security/common"
-	secconfig "github.com/DataDog/datadog-agent/pkg/security/config"
 	"github.com/DataDog/datadog-agent/pkg/security/proto/api"
 	"github.com/DataDog/datadog-agent/pkg/security/seclog"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/storage/backend"
 	grpcutils "github.com/DataDog/datadog-agent/pkg/security/utils/grpc"
+	"github.com/DataDog/datadog-agent/pkg/util/system/socket"
 )
 
 // RuntimeSecurityAgent represents the main wrapper for the Runtime Security product
@@ -310,7 +310,7 @@ func (rsa *RuntimeSecurityAgent) setupGPRC() error {
 			return errors.New("runtime_security_config.socket must be set")
 		}
 
-		family, socketPath := secconfig.GetSocketAddress(socketPath)
+		family, socketPath := socket.GetSocketAddress(socketPath)
 		if family == "unix" && runtime.GOOS == "windows" {
 			return fmt.Errorf("unix sockets are not supported on Windows")
 		}
